take one at removing yamlprojectconfig. running into weird test issue, going to repro in fresh env
